pyproject.toml
[tool.poetry]
name = "fastapi-09112023"
version = "0.1.0"
description = ""
authors = ["Василь Картичак <w.i.k.mailua@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
fastapi = {extras = ["all"], version = "^0.105.0"}
python-dotenv = "^1.0.0"
uvicorn = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
